% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_plots.R
\name{athlete_colours}
\alias{athlete_colours}
\title{Assign a colour to each paddler within their crew, steerer last}
\usage{
athlete_colours(records)
}
\description{
Assign a colour to each paddler within their crew, steerer last
}
\keyword{internal}
