% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_detect.R
\name{connected_components}
\alias{connected_components}
\title{Connected components of an undirected graph given as edge endpoints}
\usage{
connected_components(from, to)
}
\description{
Connected components of an undirected graph given as edge endpoints
}
\keyword{internal}
