% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{download_single_exercise_internal}
\alias{download_single_exercise_internal}
\title{Helper function to download exercises for a single athlete}
\usage{
download_single_exercise_internal(session, athlete_id, team_id)
}
\arguments{
\item{session}{A treinus_session object}

\item{athlete_id}{Integer ID of the athlete}

\item{team_id}{Integer ID of the team}
}
\value{
A tibble with exercise data for the athlete
}
\description{
Downloads exercises for a single athlete from the API.
}
\keyword{internal}
