% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_detect.R
\name{hms_seconds}
\alias{hms_seconds}
\title{Seconds since midnight from an "HH:MM:SS" string}
\usage{
hms_seconds(x)
}
\description{
Seconds since midnight from an "HH:MM:SS" string
}
\keyword{internal}
