% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_quality.R
\name{hr_never_acquired}
\alias{hr_never_acquired}
\title{Heart rate implausibly low while the boat is working}
\usage{
hr_never_acquired(records, boat_speed, th)
}
\description{
Heart rate implausibly low while the boat is working
}
\keyword{internal}
