% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{infer_sql_type}
\alias{infer_sql_type}
\title{Infer SQLite column type from an R vector}
\usage{
infer_sql_type(x)
}
\arguments{
\item{x}{An R vector}
}
\value{
A character string with the SQLite type name
}
\description{
Infer SQLite column type from an R vector
}
\keyword{internal}
