% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_quality.R
\name{longest_flat_run}
\alias{longest_flat_run}
\title{Longest stretch, in seconds, over which a value never changed}
\usage{
longest_flat_run(x, ts)
}
\description{
Longest stretch, in seconds, over which a value never changed
}
\keyword{internal}
