% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_plots.R
\name{theme_treinus}
\alias{theme_treinus}
\title{A plain theme for session figures}
\usage{
theme_treinus(base_size = 11)
}
\arguments{
\item{base_size}{Base font size.}
}
\value{
A ggplot2 theme.
}
\description{
A plain theme for session figures
}
