% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_measure.R
\name{treinus_crew_gap}
\alias{treinus_crew_gap}
\title{Signed speed difference between two crews}
\usage{
treinus_crew_gap(boat_speed, a, b)
}
\arguments{
\item{boat_speed}{Output of \code{\link[=treinus_boat_speed]{treinus_boat_speed()}}.}

\item{a, b}{Crew names. Positive \code{dif} means \code{a} is faster.}
}
\value{
Bin, both speeds and their difference.
}
\description{
Signed speed difference between two crews
}
