% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{treinus_db_info}
\alias{treinus_db_info}
\title{Get database information}
\usage{
treinus_db_info()
}
\value{
A tibble with database statistics
}
\description{
Returns information about the exercise database.
}
\examples{
\dontrun{
db_info <- treinus_db_info()
}

}
