% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{treinus_db_path}
\alias{treinus_db_path}
\title{Get the database file path}
\usage{
treinus_db_path()
}
\value{
Character string with the database file path.
}
\description{
Returns the path to the rtreinus database file.
}
\examples{
treinus_db_path()

}
