% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{treinus_get_exercises_db}
\alias{treinus_get_exercises_db}
\title{Get exercises database table (lazy)}
\usage{
treinus_get_exercises_db()
}
\value{
exercises
}
\description{
Get exercises database table (lazy)
}
