% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_plots.R
\name{treinus_plot_speed}
\alias{treinus_plot_speed}
\title{Speed of each crew over time}
\usage{
treinus_plot_speed(
  boat_speed,
  title = "Velocidade das canoas",
  subtitle = paste("Mediana da tripulação,", "média móvel de 60 s")
)
}
\arguments{
\item{boat_speed}{Output of \code{\link[=treinus_boat_speed]{treinus_boat_speed()}}.}

\item{title, subtitle}{Plot text.}
}
\value{
A ggplot object.
}
\description{
Speed of each crew over time
}
