% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_plots.R
\name{treinus_plot_steerer_cadence}
\alias{treinus_plot_steerer_cadence}
\title{Cadence of the steerer against the crew's stroke line}
\usage{
treinus_plot_steerer_cadence(cadence_line)
}
\arguments{
\item{cadence_line}{Output of \code{\link[=treinus_cadence_line]{treinus_cadence_line()}}.}
}
\value{
A ggplot object.
}
\description{
Cadence of the steerer against the crew's stroke line
}
