% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_settings.R
\name{treinus_write_settings}
\alias{treinus_write_settings}
\title{Write a session settings file}
\usage{
treinus_write_settings(settings, path)
}
\arguments{
\item{settings}{A settings list.}

\item{path}{Destination path.}
}
\value{
\code{path}, invisibly.
}
\description{
Write a session settings file
}
