% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{treinus_cache_dir}
\alias{treinus_cache_dir}
\title{Get the cache directory path}
\usage{
treinus_cache_dir()
}
\value{
Character string with the cache directory path.
}
\description{
Returns the path to the rtreinus cache directory, following
platform-specific conventions via \code{\link[tools:userdir]{tools::R_user_dir()}}.
}
\examples{
treinus_cache_dir()

}
\seealso{
\code{\link[=treinus_clear_cache]{treinus_clear_cache()}}, \code{\link[=treinus_cache_info]{treinus_cache_info()}}
}
